pyproject.toml
[project]
name = "knowledge-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Olaf Geibig", email = "olaf.geibig@gmail.com" },
]
keywords = ["knowledge", "mcp", "cli", "search", "fastmcp", "lightrag"]

dependencies = [
    "fastmcp>=2.2.1",
    "graspologic>=3.4.1",
    "lightrag-hku[api]>=1.3.2",
    "nano-vectordb>=0.0.4.3",
    "networkx>=3.4.2",
    "openai>=1.75.0",
    "pipmaster>=0.5.4",
    "pydantic>=2.11.3",
    "pyyaml>=6.0",
    "tenacity>=9.1.2",
    "textract>=1.6.5",
    "tiktoken>=0.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[project.urls]
Homepage = "https://github.com/olafgeibig/knowledge-mcp"
Repository = "https://github.com/olafgeibig/knowledge-mcp"

[project.scripts]
knowledge-mcp = "knowledge_mcp.cli:main"
test = "pytest:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["knowledge_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
